Extracting the S3 Client logic from the S3QueryService This will allow me to utilize that logic in a seperate class for moving files
